Some explanation is necessary here.
I'm decoupling validation from population in order to make this thing more OO
friendly. The validation broke in the process of factoring it out, but it
actually looks like it's producing better code this way. So I'm going to
temporarily comment out the validation step so that the test suite runs
as usual, at least for now.

I can do this just before 'release', so it's not going to destabilize code.