Commented out some zombie code
Not quite dead, I expect some of it will be reused as I get more test working.
Hence the reason I want to keep it around. I know that source code control
means that I can revitalize almost any of this code, but I'd just as soon not
have to remember to dig through the .attic for things.