First half of the here-doc fix
Capture the text between EOL and the here-doc marker - Doesn't quite do that
correctly because multiple here-docs can start on a single line, but this
takes care of that 80% for the moment.

Then store the here-doc start and end in the object for use in the final pass.

In the final pass (not quite where it should be), check text gaps to see if
they're actually un-captured here-docs. If so, fill the gap with a
Sir-Not-Appearing-In-This-Statement token which will *not* be present when
you, dear reader, are iterating over this structure.